update user_id from long to integer to match db schema

+++ backend/src/main/java/com/roc/app/user/general/UserService.java
@@ -14,7 +14,7 @@ public class UserService implements UserDetailsService {
         this.userRepository = userRepository;
     }
 
-    public User getUserByUserId(Long userId) {
+    public User getUserByUserId(Integer userId) {
         return userRepository.findById(userId).orElseThrow(() -> new UserNotFoundException(userId));
     }
